# backend/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Partial_IUser_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IIngredient": {
				"properties": {
					"name": {
						"type": "string"
					},
					"quantity": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"quantity"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IStep": {
				"properties": {
					"stepNumber": {
						"type": "number",
						"format": "double"
					},
					"instruction": {
						"type": "string"
					}
				},
				"required": [
					"stepNumber",
					"instruction"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IRecipes": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"ingredients": {
						"items": {
							"$ref": "#/components/schemas/IIngredient"
						},
						"type": "array"
					},
					"steps": {
						"items": {
							"$ref": "#/components/schemas/IStep"
						},
						"type": "array"
					},
					"cookingTime": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"updateBy": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"description",
					"ingredients",
					"steps",
					"cookingTime",
					"type",
					"image"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Partial_IRecipes_": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"ingredients": {
						"items": {
							"$ref": "#/components/schemas/IIngredient"
						},
						"type": "array"
					},
					"steps": {
						"items": {
							"$ref": "#/components/schemas/IStep"
						},
						"type": "array"
					},
					"cookingTime": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"updateBy": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IUser": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IAuth": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "recetas-de-cocina",
		"version": "1.0.0",
		"description": "App permite a los usuarios compartir recetas, buscar por ingredientes y guardar sus favoritas.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/api/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"UserController"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"UserController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"UserController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/users/{userId}": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"UserController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IUser_"
							}
						}
					}
				}
			}
		},
		"/api/recipes": {
			"get": {
				"operationId": "GetAllRecipes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"RecipesController"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateRecipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"RecipesController"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRecipes"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteRecipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"RecipesController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "_id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/recipes/{_id}": {
			"get": {
				"operationId": "GetRecipeById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"RecipesController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/recipes/{userId}": {
			"get": {
				"operationId": "GetRecipesByUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"RecipesController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/recipes/{recipeId}": {
			"put": {
				"operationId": "UpdateRecipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"RecipesController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "recipeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IRecipes_"
							}
						}
					}
				}
			}
		},
		"/api/auth/register": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"200": {
						"description": "msg whit result.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint for register New User to DB.",
				"tags": [
					"AuthController"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "data for register.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUser",
								"description": "data for register."
							}
						}
					}
				}
			}
		},
		"/api/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"AuthController"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAuth"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}